manual_version/requirements.txt
pymongo
paho-mqtt
hbmqtt
